controller and route seperated
